Compilable scipp-dataset after removal of realign support
