+++ dataset/reduction.cpp
@@ -3,7 +3,6 @@
 /// @file
 /// @author Simon Heybrock
 #include "scipp/common/numeric.h"
-#include "scipp/common/overloaded.h"
 
 #include "scipp/variable/reduction.h"
 
@@ -18,15 +17,14 @@ namespace scipp::dataset {
 DataArray flatten(const DataArrayConstView &a, const Dim dim) {
   return apply_to_data_and_drop_dim(
       a,
-      overloaded{no_realigned_support,
-                 [](const auto &x, const Dim dim_, const auto &mask_) {
-                   if (!contains_events(x) && min(x, dim_) != max(x, dim_))
-                     throw except::EventDataError(
-                         "flatten with non-constant scalar weights not "
-                         "possible yet.");
-                   return contains_events(x) ? flatten(x, dim_, mask_)
-                                             : copy(x.slice({dim_, 0}));
-                 }},
+      [](const auto &x, const Dim dim_, const auto &mask_) {
+        if (!contains_events(x) && min(x, dim_) != max(x, dim_))
+          throw except::EventDataError(
+              "flatten with non-constant scalar weights not "
+              "possible yet.");
+        return contains_events(x) ? flatten(x, dim_, mask_)
+                                  : copy(x.slice({dim_, 0}));
+      },
       dim, a.masks());
 }
 
@@ -34,15 +32,6 @@ Dataset flatten(const DatasetConstView &d, const Dim dim) {
   return apply_to_items(
       d, [](auto &&... _) { return flatten(_...); }, dim);
 }
-
-namespace {
-UnalignedData sum(Dimensions dims, const DataArrayConstView &unaligned,
-                  const Dim dim, const MasksConstView &masks) {
-  static_cast<void>(masks); // relevant masks are part of unaligned as well
-  dims.erase(dim);
-  return {dims, flatten(unaligned, dim)};
-}
-} // namespace
 
 DataArray sum(const DataArrayConstView &a, const Dim dim) {
   return apply_to_data_and_drop_dim(
@@ -60,9 +49,7 @@ Dataset sum(const DatasetConstView &d, const Dim dim) {
 
 DataArray mean(const DataArrayConstView &a, const Dim dim) {
   return apply_to_data_and_drop_dim(
-      a,
-      overloaded{no_realigned_support, [](auto &&... _) { return mean(_...); }},
-      dim, a.masks());
+      a, [](auto &&... _) { return mean(_...); }, dim, a.masks());
 }
 
 Dataset mean(const DatasetConstView &d, const Dim dim) {
